archive: make --add-virtual-file honor --prefix
The documentation for archive describes the `--add-virtual-file` option
thusly:

  The path of the file in the archive is built by concatenating the
  value of the last `--prefix` moption (if any) before this
  `--add-virtual-file` and <path>.

The `--add-file` documentation is similar:

  The path of the file in the archive is built by concatenating the
  value of the last --prefix option (if any) before this --add-file and
  the basename of <file>.

Notably both explicitly state that they honor the last `--prefix` option
before the `--add` option in question.

The implementation of `--add-file` has always honored prefix, while
the implementation of `--add-virtual-file` has always ignored it.

Also note that `--add-virtual-file` explicitly states it will use
the full path given, while `--add-file` uses the basename of the
path it is given, which may be the reason why the design to ignore
the prefix to avoid adding it before the name that is a full path
already.

Modify archive.c to include the prefix in the path used by
`--add-virtual-file` and add checks into
the existing add-virtual-file test to verify:

* that `--prefix` is honored
* that leading path components are preserved
* that both work together and separately

Signed-off-by: Tom Scogland <scogland1@llnl.gov>
Signed-off-by: Junio C Hamano <gitster@pobox.com>