rev-list: refactor --missing=<missing-action>
Both `git rev-list` and `git pack-objects` support a
`--missing=<missing-action>` feature, but they currently don't share
any code for that.

Refactor the support for `--missing=<missing-action>` in
"builtin/rev-list.c" into new "missing.{c,h}" files. In a following
commit, that refactored code will be used in "builtin/pack-objects.c"
too.

In yet a following commit, we are going to add support for a similar
'missing-action' feature to another command, and we are also going to
reuse code from the new "missing.{c,h}" files.

As `enum missing_action` and parse_missing_action_value() are moved to
"missing.{c,h}", we need to modify the latter a bit, so that it stops
updating any global variable, but instead returns the parsed value or
-1 on error.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>